aider: Fixed book thumbnail image not displaying by setting the src attribute to use the coverImagePath property from each book object.
# Aider chat conversation:

USER: OK next, it seems like the img for the book thumbnail doesn't have a src. The src should point to the `coverImagePath` so that the thumbnail is displayed.

ASSISTANT: To fix the issue with the book thumbnail image not displaying, we need to ensure that the `src` attribute of the `img` tag is correctly set to use the `coverImagePath` property from each book object. Here's the diff for the necessary change in the `BookLibrary.js` file:

```diff
--- client/src/components/BookLibrary.js
+++ client/src/components/BookLibrary.js
@@ ... @@
 {books.length > 0 ? books.map((book) => (
   <Grid item key={book.id} xs={12} sm={6} md={4} lg={3}>
     <Paper elevation={3} sx={{ p: 2 }}>
-      <img src={book.thumbnailUrl} alt={book.title} style={{ width: '100%', height: 'auto' }} />
+      <img src={book.coverImagePath} alt={book.title} style={{ width: '100%', height: 'auto' }} />
       <Box sx={{ mt: 2, display: 'flex', justifyContent: 'space-between' }}>
         <Button variant="outlined" color="primary">
           View
         </Button>
         <Button variant="outlined" color="secondary">
           Delete
         </Button>
       </Box>
     </Paper>
   </Grid>
 )) : <p>No books available.</p>}
```

This change will ensure that the `src` attribute of the image is correctly pointing to the `coverImagePath` of the book, which should display the thumbnail image as expected.